+++ Vault/src/utopia/vault/nosql/view/ViewManyByIntIds.scala
@@ -1,8 +1,7 @@
 package utopia.vault.nosql.view
 
-import utopia.flow.collection.immutable.IntSet
-import utopia.flow.generic.casting.ValueConversions._
 import utopia.vault.nosql.template.Indexed
+import utopia.vault.sql.Condition
 
 /**
   * Common trait for view factories that provide views to multiple items at once
@@ -17,19 +16,5 @@ trait ViewManyByIntIds[+V] extends ViewFactory[V] with Indexed
 	  * @param ids Ids of the targeted items
 	  * @return Access to those items
 	  */
-	def apply(ids: IterableOnce[Int]): V = {
-		val condition = ids match {
-			case s: IntSet => index.in(s)
-			case i: Set[Int] =>
-				val kn = ids.knownSize
-				// Case: Short collection => Won't bother converting to an int set
-				if (kn >= 0 && kn <= 6)
-					index.in(i)
-				// Case: Longer collection => Converts targeted ids to an int set
-				else
-					index.in(IntSet.from(i))
-			case i => index.in(IntSet.from(i))
-		}
-		apply(condition)
-	}
+	def apply(ids: IterableOnce[Int]): V = apply(Condition.indexIn(index, ids))
 }